requirements.txt
flask==2.3.3
flask-login==0.6.2
werkzeug==2.3.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
supabase==2.0.2
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.5.0
